+++ src/main/java/com/rmj/guanzongroup/sidebarmenus/controller/PurchaseOrder_ConfirmationMonarchFoodController.java
@@ -381,23 +381,23 @@ public class PurchaseOrder_ConfirmationMonarchFoodController implements Initiali
                     ShowMessageFX.Information((String) poJSON.get("message"), psFormName, null);
                     poJSON = poPurchasingController.PurchaseOrder().OpenTransaction(poPurchasingController.PurchaseOrder().Master().getTransactionNo());
                     // Confirmation Prompt
-                    if ("error".equals(poJSON.get("result"))){
+                    if ("error".equals(poJSON.get("result"))) {
                         ShowMessageFX.Information((String) poJSON.get("message"), psFormName, null);
                         return;
-                    } 
-                            
-                    if( poPurchasingController.PurchaseOrder().Master().getTransactionStatus().equals(PurchaseOrderStatus.OPEN)
+                    }
+
+                    if (poPurchasingController.PurchaseOrder().Master().getTransactionStatus().equals(PurchaseOrderStatus.OPEN)
                             && ShowMessageFX.YesNo(null, psFormName, "Do you want to confirm this transaction?")) {
                         if ("error".equals(poJSON.get("result"))) {
                             ShowMessageFX.Information((String) poJSON.get("message"), psFormName, null);
                             return;
                         }
-                    }    
-                        
+                    }
+
                     poJSON = poPurchasingController.PurchaseOrder().ConfirmTransaction("Confirm");
-                    if ("error".equals(poJSON.get("result"))){
-                       ShowMessageFX.Information((String) poJSON.get("message"), psFormName, null);
-                       return;
+                    if ("error".equals(poJSON.get("result"))) {
+                        ShowMessageFX.Information((String) poJSON.get("message"), psFormName, null);
+                        return;
                     }
                     ShowMessageFX.Information((String) poJSON.get("message"), psFormName, null);
 
@@ -1203,7 +1203,7 @@ public class PurchaseOrder_ConfirmationMonarchFoodController implements Initiali
 
                     for (int lnCtr = 0; lnCtr < detailCount; lnCtr++) {
                         Model_PO_Detail orderDetail = poPurchasingController.PurchaseOrder().Detail(lnCtr);
-                        double lnTotalAmount = orderDetail.getUnitPrice().doubleValue() * orderDetail.getQuantity().doubleValue();
+                        double lnTotalAmount = orderDetail.getUnitPrice().doubleValue() * orderDetail.getQuantity().intValue();
                         grandTotalAmount += lnTotalAmount;
                         double lnRequestQuantity = 0.00;
                         String status = "0";
